restructured common classes
