+++ BlazorWebSocket/Pages/Login.razor.cs
@@ -19,10 +19,10 @@
 using System;
 using System.Collections.Generic;
 using System.Net.Http;
-using System.Text.Json;
 using System.Threading.Tasks;
 using System.Web;
 using Blazored.LocalStorage;
+using BlazorWebSocket.Common;
 using Microsoft.AspNetCore.Components;
 
 namespace BlazorWebSocket.Pages {
